[ #1 ] Basic multi-screen support
Using winman_getsize we grab the width, height and x and y offsets
of the screen the window is on.

We refactor winman by defining important points on the screen rather
than recomputing them again and again. This makes it easier to make
sure that the x and y offsets are always taken into account.

This provides basic multi-screen support.

However the current screen detection is a bit too brittle: a window
slightly too big to fit in its screen may be detected as being part
of another screen. Thus leading to a dramatic move + resizing
towards the "main" screen (the one at offset (0,0)).